requirements.txt
Django==5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-jose==3.3.0
python-dotenv==1.0.0
Pillow==10.1.0
numpy==1.26.2
pandas==2.1.4
scikit-learn==1.3.2
